Basically completed the front sections
All front-matter sections are now drafted except for the Use Cases and Exceptions.
I also set up the section numbering for the entire document.
I had some questions about the newer sections I had to draft from scratch:
In Section 3, is it a rationale for the standards or for Uptane? Also, are the design requirements for Uptane or for the document? This question emerged because the excellent sample standard that were forwarded to me seemed to be writing a rationale for the standard itself. I chose a middle path and si I may need to re-write.