Make delgations apply to filenames+hardware identifiers
In order for a delegation to only apply to a restricted set of ECUs, the
constraints that apply in the delegation must align with the contraints that
the ECU uses when deciding whether to install an update. Before this change the
delegation applied to filenames, but the ECU didn't care what the filename was.
Now check both the hardware identifier and the filename when deciding whether
to install an image from the director, and when deciding if a delegation
applies or not.