Abort image search if role cannot be found or verified
Addresses Issue #35, item 3.

It must be made very clear that if at any point in a traversal, a metadata
role cannot be found or verified, the traversal must be aborted entirely and
the target not found, whether or not the delegation traversed is terminating.
To not require this is to allow a targeted DoS attack to choose among
possible options in a prioritized list of delegations (or competing groups of
roles in a multi-role delegation).

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>