FV procedure: remove efficiency step, clarify nested loop
1. In the prior commits, I added a step that allowed an ECU to skip
obtaining an image if that image's release counter was less than
that of the currently-installed image.  This is just an efficiency
step, and can be accomplished a few different ways, and since it
makes this already-complex passage a bit more complicated, I'm
just going to remove it.

2. I added a few words to make the nature of the two for loops
(one nested) happening in step 6 clearer.

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>