Remove references to time server in the body of the standard
The information about the time server will be moved to deployment
considerations, as discussed in the standards meeting on 10/04/2019.

This commit replaces references to the time server with instructions
to load the current time, or the most recent securely attested time.
It may be worthwhile to add an explanatory paragraph about the need
for a secure source of time.