Addressing Issue #116 and one item in issue #119
I've now gone through and have made sure  instances of proper names for roles, ECUs, and metadata are capitalized. I also fixed some inconsistent capitalization in headings and subheads. Now only the first word in headings is capitalized, except where proper nouns appear. This addresses  Issue #116.

I also addressed the one element in Issue #119. That is I've included the missing definition of "metadata." For this, I followed Ira's suggestion and included both the NIST definition and the version proposed by @mcvand in discussion on the issue.
I also added definitions for three acronyms used in the document that were never defined.
Lastly I fixed a handful of sentences that to my eye had structural.
I was going to add the citations to our definitions at the beginning of the document, per @iramcdonald. But I don't know how to set up the citations in Markdown using this particular format. If someone can let me know how, I can take care of that, but probably not until early next week. If another team member could step in and do this, I would likely be faster.