Updated README file
This PR amends the README for the Standards repo by moving the standards work to present tense, adding the Standards title, and updating the list of publications. This will likely require another pass, as I believe Trishank has fixed some of the document links on the website that may need changing here.