adding stipulation about changes on the repo
At the July 20 Standards meeting it was agreed that changes can be made as needed in the Standard, and do not have to wait a version is released. To clarify that the repository version of the Standard will sometimes contain different content from the numbered releases, I added some wording at the top of this page. Suggestions/modifications welcomed.

We will need to update some of the other text here, including the "existing documentation." Will we add V.1.2.0 above V.1.1.0 or delete 1.1.0?