Partial response to Issue #174
Issue #174 addresses the broad issue of whether private keys must be unique. This PR addresses just one instance singled out by @allen-cain-tm  in Section 5.4.1. It adds just one sentence to separate the idea of private keys used for decryption from private keys used for signing. This does NOT address the broader issue and so it does not close out #174