Completed V.2.0.0
While doing this review, I found a bunch of "MUSTs" that had not been changed to "SHALLS." They make up the majority of changes here. Once I commit these changes to the branch, I will do another pass scanning just for MUSTs.

I have a question --should references to "nonces or counters" be removed? Line 697 , when defining a payload for the ECU version, lists "A nonce or counter to prevent a replay." I believe this only applied to the Uptane Time Server, but I didn't want to remove it till this assumption was confirmed.