replacing "compliant" with "conformant" (#238)
Following a discussion at the 7/19 Uptane Standards meeting, it was decided we should not be using the term "Uptane-compliant." As a general rule of thumb "compliance" is reserved for regulations. So, I've removed the seven mentions of the phrase and replaced it with the term "Uptane-conformant."

We may also want to add text to discuss the issue further at some point.