modifying the use of the word conformant
Per our discussion at the 8/2/22 Uptane Standards meeting, I have added a bit of text to the introduction to define the word conformant. I was going to elaborate a bit more on this topic, I think this is probably sufficient. Note that I wasn't sure how to site the sources of the definition as it is mentioned in https://www.nist.gov/itl/ssd/information-systems-group/what-thing-called-conformance but the actual citation references the ISO guide...which is behind a pay wall.