Changing SHALL to a SHOULD (#246)
Per the decision made at the 12/6 Uptane Standard Meeting, we have relaxed the requirement that verification of Targets metadata should be stopped and considered complete if the Directory repository indicates that there are no new targets. This PR reflects that change and includes the rationale for the change. I'm unsure if the rationale belongs there, so please review and let me know.